bugfix circular imports
