+++ src/pint/residuals.py
@@ -19,7 +19,6 @@ from scipy.linalg import cho_factor, cho_solve, LinAlgError
 from scipy.stats import kstest
 
 from pint import dmu
-from pint.fitter import get_phiinv
 from pint.models.dispersion_model import Dispersion
 from pint.models.parameter import maskParameter
 from pint.models.timing_model import TimingModel
@@ -31,6 +30,7 @@ from pint.utils import (
     weighted_mean,
     woodbury_dot,
     anderson_darling,
+    get_phiinv,
 )
 
 __all__ = [
